🔄 synced local 'skyvern/' with remote 'skyvern/'
Separating interfaces and implementation for browser stuff preparing for tests.

Simply moving code around unmodified (identical on text level).

Pure refactoring, no functional changes.
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Refactor to separate `BrowserState` and `BrowserManager` protocols from their implementations, organizing code into distinct files without changing functionality.
>
>   - **Refactoring**:
>     - Split `BrowserState` protocol and its implementation into `browser_state.py` and `real_browser_state.py`.
>     - Split `BrowserManager` protocol and its implementation into `browser_manager.py` and `real_browser_manager.py`.
>   - **File Changes**:
>     - Moved `BrowserState` class from `browser_factory.py` to `browser_state.py`.
>     - Moved `BrowserArtifacts` and `VideoArtifact` classes to `browser_artifacts.py`.
>     - Updated imports in `agent_functions.py`, `persistent_sessions_manager.py`, `special_browsers.py`, `replay_task_actions.py`, `script_skyvern_page.py`, `agent.py`, `agent_functions.py`, `forge_app.py`, `block.py`, `service.py`, `task_v2_service.py`, and `scraper.py` to reflect new file structure.
>   - **Behavior**:
>     - No functional changes, purely code organization for better separation of concerns.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for dad7d205db28b736d8130224c3edbbd49ff2a558. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
