+++ skyvern/forge/agent_functions.py
@@ -23,7 +23,7 @@ from skyvern.forge.sdk.workflow.models.block import BlockTypeVar
 from skyvern.services import workflow_script_service
 from skyvern.webeye.actions.action_types import POST_ACTION_EXECUTION_ACTION_TYPES
 from skyvern.webeye.actions.actions import Action
-from skyvern.webeye.browser_factory import BrowserState
+from skyvern.webeye.browser_state import BrowserState
 from skyvern.webeye.scraper.scraper import ELEMENT_NODE_ATTRIBUTES, CleanupElementTreeFunc, json_to_html
 from skyvern.webeye.utils.dom import SkyvernElement
 from skyvern.webeye.utils.page import SkyvernFrame
